+++ rattler433/CHANGELOG.md
@@ -2,5 +2,4 @@
 
 ## 🎉 New Feature
 
-- Add option to disable internal rtl_433 instance, for users running rtl_433 on a separate device.
-- Update to use AppArmor and CodeNotary Signing
+- Add Ecowitt WH51 Soil Moisture Sensor support!